Extract thread enumeration into service/threads.py, split its tests
Review feedback on #366:

- Move the head-checkpoint constants next to the code that uses them, and cut the
  comments down to the non-obvious part.
- Extract thread enumeration out of service.py into service/threads.py, alongside
  agui.py and utils.py. The route stays in service.py with the other routes; only
  the checkpointer mechanics move.
- README documented these endpoints in detail where no other endpoint is. Trim to a
  pointer, and move the authz caveat and back-compat note into the endpoint docstring
  so they surface in /redoc, where this repo documents its API.
- Split the fake-checkpointer tests out of test_service.py into test_threads.py, and
  drop the ones the real-SQLite suite covers more faithfully (filtering, ordering,
  titles, limit, dedupe, functional-API history). What's left is what a real
  checkpointer can't be made to do on demand: the row/page caps, a checkpointer that
  ignores the filter, a missing timestamp, a failing query. The fake no longer models
  the __previous__/__end__ channels at all.
